+++ ros/src/airsim_ros_pkgs/include/airsim_ros_wrapper.h
@@ -144,6 +144,7 @@ private:
     {
         msr::airlib::SensorBase::SensorType sensor_type;
         std::string sensor_name;
+        std::string data_frame;
         ros::Publisher publisher;
     };
     // utility struct for a SINGLE robot
@@ -281,7 +282,10 @@ private:
     sensor_msgs::Imu get_imu_msg_from_airsim(const msr::airlib::ImuBase::Output& imu_data) const;
     airsim_ros_pkgs::Altimeter get_altimeter_msg_from_airsim(const msr::airlib::BarometerBase::Output& alt_data) const;
     sensor_msgs::Range get_range_from_airsim(const msr::airlib::DistanceSensorData& dist_data) const;
-    sensor_msgs::PointCloud2 get_lidar_msg_from_airsim(const msr::airlib::LidarData& lidar_data, const std::string& vehicle_name) const;
+    sensor_msgs::PointCloud2 get_lidar_msg_from_airsim(const msr::airlib::LidarData& lidar_data,
+                                                       const std::string& vehicle_name,
+                                                       const std::string& sensor_name,
+                                                       const std::string& data_frame) const;
     sensor_msgs::NavSatFix get_gps_msg_from_airsim(const msr::airlib::GpsBase::Output& gps_data) const;
     sensor_msgs::MagneticField get_mag_msg_from_airsim(const msr::airlib::MagnetometerBase::Output& mag_data) const;
     airsim_ros_pkgs::Environment get_environment_msg_from_airsim(const msr::airlib::Environment::State& env_data) const;